# nsys-jax: optimise data loading and .zip creation