… .github/container/nsys_jax/nsys_jax/visualization.py
… .github/container/nsys_jax/tests/test_df_helpers.py
… .github/workflows/_ci.yaml