Some rough measurements on vanilla `jax-nccl-test` and 8xH100:

> Profile collection, whole execution: 52s (nsys), 58s (nsys-jax with this PR), 1m5s (nsys-jax without this PR)
> Profile collection, restricted range: 46s (nsys), 50s (nsys-jax with this PR), 55s (nsys-jax without this PR)
> Communication analysis, whole execution: 1.1s (with this PR), 2.1s (without this PR)
> Communication analysis, restricted range: 1.0s (with this PR), 1.7s (without this PR)

The differences are more pronounced on larger workloads with more activity.

The two bigger changes are:
- Convert `.csv` to `.parquet` as part of `nsys-jax` to avoid compressing `.csv` with Python's `lzma` module, which is slow and single-threaded. This speeds up `nsys-jax` *and* subsequent data-loading.
- A new algorithm for calculating the hidden/exposed time of communication kernels when loading profile data -- essentially this adds a fast pandas-friendly pass to identify [most] non-overlapping kernels and skip running the [relatively slow and pandas-unfriendly] overlap calculation on them. This also removes an assumption that there is no compute-compute overlap.

Otherwise there are some tweaks to pandas usage and minor reorganisations to make Python profiles more informative, and minor bugfixes in the example Jupyter notebook.