@gspschmid 2025-01-15
Here's what a single sweep over all the comm/compute intervals might look like: https://gist.github.com/gspschmid/478b056b42a5c81e00999617b18d5b71